Rename data validation function in stock_processing.py
