+++ excel_analysis/stock_processing.py
@@ -16,7 +16,7 @@ from excel_analysis.constants import (
 )
 from excel_analysis.utils.argument_parser import parse_argumentos
 from excel_analysis.utils.cargadores_de_datos import validar_y_cargar_hojas
-from excel_analysis.utils.data_validation import validar_datos_hoja
+from excel_analysis.utils.validacion_de_datos import validar_datos_hoja
 from excel_analysis.utils.display_results import almacenar_y_mostrar_resultados
 from excel_analysis.utils.entrenamiento import entrenar_y_predecir
 from excel_analysis.utils.grading_system import (
